requirements.txt
PyYAML==3.11
SQLAlchemy==0.9.6
coverage==3.7.1
nose==1.3.3
psycopg2==2.5.3
python-dateutil==2.2
six==1.7.3
unicodecsv==0.9.4
wsgiref==0.1.2
xlrd==0.9.3
